FISH-186 Local and remote command for listing store entries
