+++ appserver/extras/certificate-management/certificate-management-cli/src/main/java/fish/payara/certificate/management/cli/ListTrustStoreEntriesCommand.java
@@ -1,0 +1,87 @@
+/*
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
+ *
+ * Copyright (c) 2020 Payara Foundation and/or its affiliates. All rights reserved.
+ *
+ * The contents of this file are subject to the terms of either the GNU
+ * General Public License Version 2 only ("GPL") or the Common Development
+ * and Distribution License("CDDL") (collectively, the "License").  You
+ * may not use this file except in compliance with the License.  You can
+ * obtain a copy of the License at
+ * https://github.com/payara/Payara/blob/master/LICENSE.txt
+ * See the License for the specific
+ * language governing permissions and limitations under the License.
+ *
+ * When distributing the software, include this License Header Notice in each
+ * file and include the License file at glassfish/legal/LICENSE.txt.
+ *
+ * GPL Classpath Exception:
+ * The Payara Foundation designates this particular file as subject to the "Classpath"
+ * exception as provided by the Payara Foundation in the GPL Version 2 section of the License
+ * file that accompanied this code.
+ *
+ * Modifications:
+ * If applicable, add the following below the License Header, with the fields
+ * enclosed by brackets [] replaced by your own identifying information:
+ * "Portions Copyright [year] [name of copyright owner]"
+ *
+ * Contributor(s):
+ * If you wish your version of this file to be governed by only the CDDL or
+ * only the GPL Version 2, indicate your decision by adding "[Contributor]
+ * elects to include this software in this distribution under the [CDDL or GPL
+ * Version 2] license."  If you don't indicate a single choice of license, a
+ * recipient has the option to distribute your version of this file under
+ * either the CDDL, the GPL Version 2 or to extend the choice of license to
+ * its licensees as provided above.  However, if you add GPL Version 2 code
+ * and therefore, elected the GPL Version 2 license, then the option applies
+ * only if the new code is made subject to such option by the copyright
+ * holder.
+ */
+package fish.payara.certificate.management.cli;
+
+import com.sun.enterprise.admin.cli.CLIConstants;
+import com.sun.enterprise.admin.cli.Environment;
+import com.sun.enterprise.admin.cli.ProgramOptions;
+import com.sun.enterprise.util.SystemPropertyConstants;
+import org.glassfish.api.admin.CommandException;
+import org.glassfish.hk2.api.PerLookup;
+import org.jvnet.hk2.annotations.Service;
+
+@Service(name = "list-truststore-entries")
+@PerLookup
+public class ListTrustStoreEntriesCommand extends AbstractListStoreEntriesCommand {
+
+    @Override
+    protected int executeCommand() throws CommandException {
+        // If we're targetting an instance that isn't the DAS, use a different command
+        if (target != null && !target.equals(SystemPropertyConstants.DAS_SERVER_NAME)) {
+            ListTrustStoreEntriesLocalInstanceCommand localInstanceCommand =
+                    new ListTrustStoreEntriesLocalInstanceCommand(programOpts, env);
+            localInstanceCommand.validate();
+            return localInstanceCommand.executeCommand();
+        }
+
+        parseTrustStore();
+
+        listEntries(truststore, truststorePassword);
+
+        return CLIConstants.SUCCESS;
+    }
+
+
+
+    private class ListTrustStoreEntriesLocalInstanceCommand extends AbstractLocalInstanceListStoreEntriesCommand {
+
+        public ListTrustStoreEntriesLocalInstanceCommand(ProgramOptions programOpts, Environment env) {
+            super(programOpts, env);
+        }
+
+        protected int executeCommand() throws CommandException {
+            parseTrustStore();
+
+            listEntries(truststore, truststorePassword);
+
+            return CLIConstants.SUCCESS;
+        }
+    }
+}
