+++ appserver/extras/certificate-management/certificate-management-cli/src/main/manpages/fish/payara/certificate/management/cli/generate-self-signed-certificate.1
