+++ appserver/extras/certificate-management/certificate-management-cli/src/main/manpages/fish/payara/certificate/management/cli/remove-from-truststore.1
