+++ appserver/extras/certificate-management/certificate-management-cli/src/test/java/fish/payara/certificate/management/GenerateSelfSignedCertificateCommandTest.java
