+++ appserver/extras/certificate-management/certificate-management-console-plugin/src/main/resources/fish/payara/admingui/certificate/management/Strings.properties
@@ -1,0 +1,2 @@
+tree.certificateManagement=Certificate Management
+certificateManagement.pageTitle=Certificate Management
